app/requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
pillow
python-multipart
streamlit
requests
numpy
scikit-learn
matplotlib
seaborn
